+++ components/card/__docs__/storybook-stories.jsx
@@ -20,10 +20,12 @@ import MediaObject from '../../media-object';
 import InlineEdit from '../../forms/input/inline';
 
 const sampleItems = [
-	{ name: 'Cloudhub' },
-	{ name: 'Cloudhub + Anypoint Connectors' },
-	{ name: 'Cloud City' },
+	{ id: '0', name: 'Cloudhub' },
+	{ id: '1', name: 'Cloudhub + Anypoint Connectors' },
+	{ id: '2', name: 'Cloud City' },
 ];
+
+let currentId = 3;
 
 class DemoCard extends React.Component {
 	static displayName = 'DemoCard';
@@ -63,7 +65,10 @@ class DemoCard extends React.Component {
 		action('add')(...rest);
 
 		this.setState({
-			items: [{ name: `New item #${shortid.generate()}` }, ...this.state.items],
+			items: [
+				{ id: currentId++, name: `New item #${shortid.generate()}` },
+				...this.state.items,
+			],
 		});
 	};
 
@@ -107,7 +112,7 @@ class DemoCard extends React.Component {
 					icon={<Icon category="standard" name="document" size="small" />}
 					empty={isEmpty ? <CardEmpty heading="No Related Items" /> : null}
 				>
-					<DataTable id="SLDSDataTableExample-1" items={items} bordered>
+					<DataTable id="SLDSDataTableExample-1" items={items}>
 						<DataTableColumn label="Opportunity Name" property="name" truncate>
 							<DataTableHighlightCell search={this.state.filter} />
 						</DataTableColumn>