+++ components/combobox/combobox.jsx
@@ -1,7 +1,9 @@
+/* eslint-disable max-lines */
 /* Copyright (c) 2015-present, salesforce.com, inc. All rights reserved */
 /* Licensed under BSD 3-Clause - see LICENSE.txt or git.io/sfdc-license */
 
 /* eslint-disable jsx-a11y/role-has-required-aria-props */
+/* eslint-disable max-lines */
 
 import React from 'react';
 import PropTypes from 'prop-types';