+++ components/data-table/index.jsx
@@ -1,3 +1,4 @@
+/* eslint-disable max-lines */
 /* Copyright (c) 2015-present, salesforce.com, inc. All rights reserved */
 /* Licensed under BSD 3-Clause - see LICENSE.txt or git.io/sfdc-license */
 
@@ -20,6 +21,11 @@ import reject from 'lodash.reject';
 // This component's `checkProps` which issues warnings to developers about properties when in development mode (similar to React's built in development tools)
 import checkProps from './check-props';
 import componentDoc from './docs.json';
+
+import {
+	canUseDOM,
+	canUseEventListeners,
+} from '../../utilities/execution-environment';
 
 // ## Children
 import DataTableCell from './cell';
@@ -113,13 +119,17 @@ class DataTable extends React.Component {
 		 */
 		columnBordered: PropTypes.bool,
 		/**
+		 * Use this to enable fixed headers and scrolling columns / rows. Appearance / behavior is consistent only if used in combination with `fixedLayout`. Since scrolling is enabled, columns are not truncated unless a width is set. Due to `overflow:hidden` elements, any dialog components will need a separate render tree (portal) such as with `menuPosition: overflowBoundaryElement` in order to break out of the container.
+		 */
+		fixedHeader: PropTypes.bool,
+		/**
+		 * Use this if you are creating an advanced table (selectable, sortable, or resizable rows). Columns widths will be truncate based on width and DOM ancestors. See `fixedHeader` to enable horizontal and vertical scrolling.
+		 */
+		fixedLayout: PropTypes.bool,
+		/**
 		 * A unique ID is needed in order to support keyboard navigation and ARIA support.
 		 */
 		id: PropTypes.string,
-		/**
-		 * Use this if you are creating an advanced table (selectable, sortable, or resizable rows)
-		 */
-		fixedLayout: PropTypes.bool,
 		/**
 		 * The collection of items to render in the table. This is an array of objects with each object having keys that correspond with the  `property` prop of each `DataTableColumn`.
 		 */
@@ -133,6 +143,10 @@ class DataTable extends React.Component {
 		 */
 		noRowHover: PropTypes.bool,
 		/**
+		 * By default this function resizes the display headers when fixedHeader is `true`, but this behavior can be overridden. Passes an event and a data object with properties `headerRefs`, an array of DOM nodes referencing the `thead th` elements and `scrollerRef`, a DOM node referencing `.slds-table_header-fixed_scroller`
+		 */
+		onFixedHeaderResize: PropTypes.func,
+		/**
 		 * This function fires when the selection of rows changes. This component passes in `event, { selection }` to the function. `selection` is an array of objects from the `items` prop.
 		 *
 		 * This used to be `onChange` which is deprecated now, so that the parameters can be consistent with other components. `onChange` passed in the selection first and the event wtihout a data object.
@@ -143,11 +157,18 @@ class DataTable extends React.Component {
 		 */
 		onSort: PropTypes.func,
 		/**
+		 * By default this function attaches/detaches listeners for window resize and tbody scrolling when fixedHeader is `true`, but this behavior can be overridden. Passes an event and a data object with an `attach` boolean property to determine whether listeners should be attached, a `resizeHandler` function property that can be called as-needed, and a `scrollerRef` DOM node property that serves as a reference to `.slds-table_header-fixed_scroller`
+		 */
+		onToggleFixedHeaderListeners: PropTypes.func,
+		/**
 		 * An array of objects of selected rows. See `items` prop for shape of objects.
 		 */
 		selection: PropTypes.array,
 		/**
-		 * Specifies a select row UX pattern. `checkbox` should be used for multiple row selection. `radio` should be limited to _required_ single row selection. This prop used to be a `boolean`, a `true` value will be considered `checkbox` for backwards compatibility
+		 * Specifies a row selection UX pattern.
+		 * * `checkbox`: Multiple row selection.
+		 * * `radio`: _Required_ single row selection.
+		 * _This prop used to be a `boolean`, a `true` value will be considered `checkbox` for backwards compatibility._
 		 */
 		selectRows: PropTypes.oneOfType([
 			PropTypes.bool,
@@ -180,11 +201,34 @@ class DataTable extends React.Component {
 	constructor(props) {
 		super(props);
 		this.generatedId = shortid.generate();
+		this.headerRefs = {
+			action: [],
+			column: [],
+			select: [],
+		};
+		this.scrollerRef = null;
 	}
 
 	componentWillMount() {
 		// `checkProps` issues warnings to developers about properties (similar to React's built in development tools)
 		checkProps(DATA_TABLE, this.props, componentDoc);
+	}
+
+	componentDidMount() {
+		if (this.props.fixedHeader) {
+			this.toggleFixedHeaderListeners(true);
+			this.resizeFixedHeaders();
+		}
+	}
+
+	componentDidUpdate() {
+		if (this.props.fixedHeader) {
+			this.resizeFixedHeaders();
+		}
+	}
+
+	componentWillUnmount() {
+		this.toggleFixedHeaderListeners(false);
 	}
 
 	getId() {
@@ -240,6 +284,70 @@ class DataTable extends React.Component {
 		}
 	};
 
+	resizeFixedHeaders = (event) => {
+		const headerRefs = [].concat(
+			this.headerRefs.select,
+			this.headerRefs.column,
+			this.headerRefs.action
+		);
+
+		if (this.props.onFixedHeaderResize) {
+			this.props.onFixedHeaderResize(event, {
+				headerRefs,
+				scrollerRef: this.scrollerRef,
+			});
+		} else if (headerRefs.length > 0) {
+			let documentScrollLeft = 0;
+
+			if (canUseDOM) {
+				documentScrollLeft = document.documentElement.scrollLeft;
+			}
+
+			headerRefs.forEach((column) => {
+				if (column && canUseDOM) {
+					const columnLeft =
+						column.getBoundingClientRect().left + documentScrollLeft;
+					let wrapperLeft = 0;
+
+					if (this.scrollerRef) {
+						wrapperLeft =
+							this.scrollerRef.getBoundingClientRect().left +
+							documentScrollLeft;
+					}
+
+					const cellFixed = column.querySelector('.slds-cell-fixed');
+					const linkFixed = cellFixed.firstChild;
+
+					if (cellFixed) {
+						cellFixed.style.left = `${columnLeft - wrapperLeft}px`;
+						linkFixed.style.width = `${column.offsetWidth}px`;
+					}
+				}
+			});
+		}
+	};
+
+	toggleFixedHeaderListeners = (attach) => {
+		if (this.props.onToggleFixedHeaderListeners) {
+			this.props.onToggleFixedHeaderListeners(
+				{},
+				{
+					attach,
+					resizeHandler: this.resizeFixedHeaders,
+					scrollerRef: this.scrollerRef,
+				}
+			);
+		} else {
+			const action = [`${attach ? 'add' : 'remove'}EventListener`];
+			if (canUseEventListeners) {
+				window[action]('resize', this.resizeFixedHeaders);
+			}
+			if (canUseEventListeners && this.scrollerRef) {
+				this.scrollerRef[action]('scroll', this.resizeFixedHeaders);
+			}
+		}
+	};
+
 	// ### Render
 	render() {
 		const ariaProps = {};
@@ -278,7 +386,16 @@ class DataTable extends React.Component {
 				child &&
 				child.type.displayName === DataTableRowActions.displayName
 			) {
-				RowActions = child;
+				const dropdown = child.props.dropdown;
+				const dropdownPropOverrides = {};
+				if (this.props.fixedHeader) {
+					dropdownPropOverrides.menuPosition = 'overflowBoundaryElement';
+				}
+				RowActions = React.cloneElement(child, {
+					dropdown: dropdown
+						? React.cloneElement(dropdown, dropdownPropOverrides)
+						: null,
+				});
 			}
 		});
 
@@ -309,13 +426,21 @@ class DataTable extends React.Component {
 			ariaProps['aria-multiselectable'] = 'true';
 		}
 
-		return (
+		// This guarantees there are never any old header references if props change
+		this.headerRefs = {
+			action: RowActions ? this.headerRefs.action : [],
+			column: this.headerRefs.column.slice(0, columns.length),
+			select: canSelectRows ? this.headerRefs.select : [],
+		};
+
+		let component = (
 			<table
 				{...ariaProps}
 				className={classNames(
 					'slds-table',
 					{
 						'slds-table_fixed-layout': this.props.fixedLayout,
+						'slds-table_header-fixed': this.props.fixedHeader,
 						'slds-table_resizable-cols': this.props.fixedLayout,
 						'slds-table_bordered': !this.props.unborderedRow,
 						'slds-table_cell-buffer':
@@ -335,6 +460,18 @@ class DataTable extends React.Component {
 				<DataTableHead
 					assistiveText={assistiveText}
 					allSelected={allSelected}
+					fixedHeader={this.props.fixedHeader}
+					headerRefs={(ref, index) => {
+						if (index === 'action' || index === 'select') {
+							if (ref) {
+								this.headerRefs[index][0] = ref;
+							} else {
+								this.headerRefs[index] = [];
+							}
+						} else {
+							this.headerRefs.column[index] = ref;
+						}
+					}}
 					indeterminateSelected={indeterminateSelected}
 					canSelectRows={canSelectRows}
 					columns={columns}
@@ -371,6 +508,30 @@ class DataTable extends React.Component {
 				</tbody>
 			</table>
 		);
+
+		if (this.props.fixedHeader) {
+			component = (
+				<div
+					className="slds-table_header-fixed_container"
+					style={{ height: '100%' }}
+				>
+					<div
+						className="slds-table_header-fixed_scroller"
+						ref={(ref) => {
+							this.scrollerRef = ref;
+						}}
+						style={{
+							height: '100%',
+							overflow: 'auto',
+						}}
+					>
+						{component}
+					</div>
+				</div>
+			);
+		}
+
+		return component;
 	}
 }