+++ components/input/__tests__/input.browser-test.jsx
@@ -1,3 +1,4 @@
+/* eslint-disable max-lines */
 /* eslint-disable react/no-render-return-value */
 import React from 'react';
 import ReactDOM from 'react-dom';